requirements.txt
requests
pandas
python-dotenv
# tensorflow
torch
torchvision
torchaudio
transformers